add hamburger menu to mobile styling
