+++ components/Nav.jsx
@@ -1,15 +1,22 @@
 'use client';
 
+import { useState } from 'react';
 import Link from 'next/link';
 import '../app/globals.css';
 import { MdBrightness4 } from 'react-icons/md';
-
+import { CgMenuRight } from 'react-icons/cg';
 
 const Nav = () => {
+    const [isMenuOpen, setIsMenuOpen] = useState(false);
+
     const toggleDarkMode = () => {
         document.body.classList.toggle('dark-mode');
         const isDarkMode = document.body.classList.contains('dark-mode');
         localStorage.setItem('darkMode', isDarkMode ? 'enabled' : 'disabled');
+    };
+
+    const toggleMenu = () => {
+        setIsMenuOpen(!isMenuOpen);
     };
 
     return (
@@ -25,7 +32,8 @@ const Nav = () => {
                         </h1>
                     </div>
                 </Link>
-                <div className='flex space-x-6'>
+
+                <div className=' hidden sm:flex space-x-6'>
                     <Link href="/faqs" passHref>
                         <div 
                             className='text-aliceblue font-semibold text-xl font-raleway'
@@ -34,7 +42,15 @@ const Nav = () => {
                             FAQs
                         </div>
                     </Link>
-                    <Link href="/contact" passHref>
+                    <Link href="/about" passHref>
+                        <div 
+                            className='text-aliceblue font-semibold text-xl font-raleway'
+                            style={{ fontFamily: 'Raleway, sans-serif' }}
+                        >
+                            About 
+                        </div>
+                    </Link> 
+                    <Link href="/about#contact" passHref>
                         <div 
                             className='text-aliceblue font-semibold text-xl font-raleway'
                             style={{ fontFamily: 'Raleway, sans-serif' }}
@@ -46,6 +62,35 @@ const Nav = () => {
                         <MdBrightness4 />
                     </button>
                 </div>
+
+                <div className='sm:hidden'>
+                <button onClick={toggleMenu} className='cursor-pointer'>
+                    <CgMenuRight className='text-2xl text-aliceblue' />
+                </button>
+            </div>
+
+                {isMenuOpen && (
+                    <div className='absolute top-0 left-0 h-full w-full bg-black bg-opacity-50 z-10 flex flex-col items-center justify-center'>
+                        <Link href="/faqs" passHref>
+                            <div className='text-aliceblue font-semibold text-xl font-raleway mb-4 cursor-pointer'>
+                                FAQs
+                            </div>
+                        </Link>
+                        <Link href="/about" passHref>
+                            <div className='text-aliceblue font-semibold text-xl font-raleway mb-4 cursor-pointer'>
+                                About
+                            </div>
+                        </Link>
+                        <Link href="/about#contact" passHref>
+                            <div className='text-aliceblue font-semibold text-xl font-raleway mb-4 cursor-pointer'>
+                                Contact
+                            </div>
+                        </Link>
+                        <button onClick={toggleDarkMode} className='cursor-pointer'>
+                            <MdBrightness4 className='text-2xl text-aliceblue' />
+                        </button>
+                    </div>
+                )}
             </div>
     );
 }
